+++ src/classification/naive-bayes.js
@@ -6,11 +6,14 @@
  *
  * [Reference]: https://en.wikipedia.org/wiki/Naive_Bayes_classifier
  */
+import {vec} from '../helpers/vectors';
 import {mat} from '../helpers/matrices';
 import {mean, variance} from '../stats/descriptive';
 
 /**
  * The classifier's class.
+ *
+ * Note: add a way to import/export the computed model (#.toJSON also).
  *
  * @constructor
  */
@@ -22,63 +25,139 @@ export default class NaiveBayes {
    * @return {NaiveBayes} - Returns itself for chaining.
    */
   reset() {
-
+    this.classes = null;
+    this.priors = null;
+    this.dimensions = 0;
+    this.theta = null;
+    this.sigma = null;
   }
 
   /**
    * Method used to train the classifier and taking the dataset's vectors &
    * labels.
    *
-   * @param  {array} X    - Training vectors.
-   * @param  {array} y    - Target values.
-   * @return {NaiveBayes} - Returns itself for chaining.
+   * Note: Decide whether to add espilon to sigma in near future.
    *
-   * @throws {Error} - Will throw if X and y are not of same length.
+   * @param  {array}      features    - Training vectors.
+   * @param  {array}      labels      - Target values.
+   * @return {NaiveBayes}             - Returns itself for chaining.
+   *
+   * @throws {Error} - Will throw if features and labels are not of same length.
    */
-  fit(X, y) {
-    const nbVectors = X.length;
+  fit(features, labels) {
+    const nbVectors = features.length;
 
-    if (nbVectors !== y.length)
+    if (nbVectors !== labels.length)
       throw Error('talisman/classification/naive-bayes: given arrays have different lengths.');
 
     // Resetting internal state
     this.reset();
 
-    // Espilon
-    const XVariances = X.map(variance),
-          maxVariance = Math.max(...XVariances);
-
-    const epsilon = 1e-9 * maxVariance;
-
     // Classes
-    const classes = {};
+    const classes = {},
+          priors = {};
 
     // Finding unique classes
-    for (let i = 0, l = y.length; i < l; i++) {
-      const label = y[i];
+    for (let i = 0, l = labels.length; i < l; i++) {
+      const label = labels[i];
 
-      classes[label] = classes[label] || {count: 0};
-      classes[label].count++;
+      classes[label] = classes[label] || 0;
+      classes[label]++;
     }
 
-    // Lengths
-    const dimensions = X[0].length,
-          nbClasses = Object.keys(classes).length;
+    for (const k in classes)
+      priors[k] = classes[k] / nbVectors;
 
-    // Building summaries
+    // Lengths
+    const dimensions = features[0].length;
+
+    // Building matrices
+    const matrices = {},
+          offsets = {},
+          featureSets = mat(dimensions, nbVectors);
+
     for (const k in classes) {
-      const c = classes[k];
-      c.matrix = mat(this.dimensions, c.count);
+      matrices[k] = mat(dimensions, classes[k]);
+      offsets[k] = vec(dimensions, 0);
     }
 
     for (let i = 0; i < nbVectors; i++) {
-      const label = y[i],
-            c = classes[label];
+      const label = labels[i],
+            matrix = matrices[label];
 
-      for (let j = 0; j < this.dimensions; j++) {
-        c.matrix[j][i] = X[i][j];
+      for (let j = 0; j < dimensions; j++) {
+        matrix[j][offsets[label][j]++] = features[i][j];
+        featureSets[j][i] = features[i][j];
       }
     }
+
+    // Epsilon
+    const maxVariance = Math.max(...featureSets.map(f => variance(f))),
+          espilon = 1e-9 * maxVariance;
+
+    // Computing means & variances
+    const theta = {},
+          sigma = {};
+
+    for (const k in matrices) {
+      theta[k] = [];
+      sigma[k] = [];
+
+      for (let i = 0; i < dimensions; i++) {
+        theta[k][i] = mean(matrices[k][i]);
+        sigma[k][i] = variance(matrices[k][i], theta[k][i]) + espilon;
+      }
+    }
+
+    this.classes = classes;
+    this.priors = priors;
+    this.dimensions = dimensions;
+    this.theta = theta;
+    this.sigma = sigma;
+
+    return this;
+  }
+
+  /**
+   * Method used to get the joint log likelihood for a new vector.
+   *
+   * @param  {array} vector - The vector to classify.
+   * @return {object}       - The probabilities.
+   *
+   * @throw {Error} - The classifier cannot predict if not yet fitted.
+   * @throw {Error} - The classifier expects a vector of correct dimension.
+   */
+  jointLogLikelihood(vector) {
+    if (!this.theta)
+      throw Error('talisman/classification/naive-bayes.probabilities: the classifier is not yet fitted');
+
+    if (vector.length !== this.dimensions)
+      throw Error(`talisman/classification/naive-bayes.probabilities: the given vector is not of correct dimension (${vector.length} instead of ${this.dimensions}).`);
+
+    const probabilities = {};
+
+    for (const k in this.classes) {
+      const theta = this.theta[k],
+            sigma = this.sigma[k],
+            jointi = Math.log(this.priors[k]);
+
+      let s1 = 0,
+          s2 = 0;
+
+      for (let i = 0; i < this.dimensions; i++) {
+        const t = theta[i],
+              s = sigma[i],
+              x = vector[i];
+
+        s1 += Math.log(2 * Math.PI * s);
+        s2 += Math.pow(x - t, 2) / s;
+      }
+
+      const nij = (-0.5 * s1) - (0.5 * s2);
+      probabilities[k] = jointi + nij;
+    }
+
+    return probabilities;
   }
 
   /**
@@ -88,7 +167,20 @@ export default class NaiveBayes {
    * @return {mixed}        - The predicted label.
    */
   predict(vector) {
+    const probabilities = this.jointLogLikelihood(vector);
 
+    // Finding the best class
+    let bestClass = null,
+        bestScore = -Infinity;
+
+    for (const k in probabilities) {
+      if (bestScore < probabilities[k]) {
+        bestClass = k;
+        bestScore = probabilities[k];
+      }
+    }
+
+    return bestClass;
   }
 }